Çekme isteği öncesindeki son değişiklikler
