--- app/src/main/java/com/crypto/katip/ProfileActivity.java
@@ -1,14 +1,0 @@
-package com.crypto.katip;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class ProfileActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_profile);
-    }
-}
