+++ app/src/main/java/com/crypto/katip/WelcomeActivity.java
@@ -20,8 +20,5 @@ public class WelcomeActivity extends AppCompatActivity {
         } else {
             startActivity(new Intent(WelcomeActivity.this, LoginActivity.class));
         }
-
-
-
     }
 }
