+++ app/src/main/java/com/crypto/katip/controllers/UserController.java
@@ -6,7 +6,7 @@ import com.crypto.katip.database.DbHelper;
 import com.crypto.katip.models.User;
 
 public class UserController {
-    private DbHelper dbHelper;
+    private final DbHelper dbHelper;
 
     public UserController(DbHelper dbHelper) {
         this.dbHelper = dbHelper;
@@ -14,20 +14,24 @@ public class UserController {
 
     public void save(String username, String password) {
         User user = new User(username, password, dbHelper);
+
         user.save();
     }
 
     public boolean isRegistered(String username, String password) {
         User user = User.getInstance(username, dbHelper);
+
         if (user != null) {
             user.setPassword(password);
             return user.isRegistered();
         }
+
         return false;
     }
 
     public void updatePassword(int id, String password) {
         User user = User.getInstance(id, dbHelper);
+
         if (user != null) {
             user.setPassword(password);
             user.update();
@@ -36,6 +40,7 @@ public class UserController {
 
     public void updateUsername(int id, String username) {
         User user = User.getInstance(id, dbHelper);
+
         if (user != null) {
             user.setUsername(username);
             user.update();
@@ -44,6 +49,7 @@ public class UserController {
 
     public void remove(int id) {
         User user = User.getInstance(id, dbHelper);
+
         if (user != null) {
             user.remove();
         }
