+++ app/src/main/java/com/crypto/katip/database/DbHelper.java
@@ -4,8 +4,6 @@ import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 
-import androidx.annotation.Nullable;
-
 public class DbHelper extends SQLiteOpenHelper {
 
     public DbHelper(Context context) {
