--- app/src/main/java/com/crypto/katip/ui/profile/ProfileViewModel.java
@@ -1,18 +1,0 @@
-package com.crypto.katip.ui.profile;
-
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.ViewModel;
-
-public class ProfileViewModel extends ViewModel {
-    private MutableLiveData<String> mText;
-
-    public ProfileViewModel() {
-        mText = new MutableLiveData<>();
-        mText.setValue("This is home fragment");
-    }
-
-    public LiveData<String> getText() {
-        return mText;
-    }
-}
