+++ app/src/main/java/com/crypto/katip/viewmodels/login/LoginViewModel.java
@@ -1,4 +1,4 @@
-package com.crypto.katip.ui.login;
+package com.crypto.katip.viewmodels.login;
 
 import android.content.Context;
 
@@ -6,10 +6,9 @@ import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
 import com.crypto.katip.controllers.LoginController;
-import com.crypto.katip.database.DbHelper;
 
 public class LoginViewModel extends ViewModel {
-    private MutableLiveData<LoginResult> result = new MutableLiveData<>();
+    private final MutableLiveData<LoginResult> result = new MutableLiveData<>();
 
     public MutableLiveData<LoginResult> getResult() {
         return result;
@@ -17,6 +16,7 @@ public class LoginViewModel extends ViewModel {
 
     public void login(String username, String password, Context context) {
         LoginController loginController = LoginController.getInstance();
+
         loginController.login(username, password, context);
         if (loginController.getUser() != null) {
             result.setValue(new LoginResult(loginController.getUser()));
