+++ app/src/main/java/com/crypto/katip/viewmodels/login/LoginViewModelFactory.java
@@ -1,16 +1,16 @@
-package com.crypto.katip.ui.register;
+package com.crypto.katip.viewmodels.login;
 
 import androidx.annotation.NonNull;
 import androidx.lifecycle.ViewModel;
 import androidx.lifecycle.ViewModelProvider;
 
-public class RegisterViewModelFactory implements ViewModelProvider.Factory {
+public class LoginViewModelFactory implements ViewModelProvider.Factory {
     @NonNull
     @Override
     @SuppressWarnings("unchecked")
     public <T extends ViewModel> T create(@NonNull Class<T> modelClass) {
-        if (modelClass.isAssignableFrom(RegisterViewModel.class)) {
-            return (T) new RegisterViewModel();
+        if(modelClass.isAssignableFrom(LoginViewModel.class)) {
+            return (T) new LoginViewModel();
         } else {
             throw new IllegalArgumentException("Unknown ViewModel class");
         }
