--- app/src/main/res/navigation/mobile_navigation.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/mobile_navigation"
-    app:startDestination="@id/navigation_profile">
-
-    <fragment
-        android:id="@+id/navigation_profile"
-        android:name="com.crypto.katip.ui.profile.ProfileFragment"
-        tools:layout="@layout/fragment_profile"
-        android:label="@string/title_profile" />
-    <fragment
-        android:id="@+id/navigation_chat"
-        android:name="com.crypto.katip.ui.chat.ChatFragment"
-        tools:layout="@layout/fragment_chat"
-        android:label="@string/title_chat" />
-</navigation>
